Move NetworkUtils outside data layer since it requires Android application context
