+++ app/src/main/java/com/example/briefly/domain/network/NetworkUtils.kt
@@ -1,0 +1,5 @@
+package com.example.briefly.domain.network
+
+interface NetworkUtils {
+    fun isConnected(): Boolean
+}
